+++ botx/__init__.py
@@ -50,6 +50,7 @@ from botx.models.messages.sending.markup import MessageMarkup
 from botx.models.messages.sending.message import SendingMessage
 from botx.models.messages.sending.options import MessageOptions, NotificationOptions
 from botx.models.messages.sending.payload import MessagePayload, UpdatePayload
+from botx.models.smartapps import SendingSmartAppEvent, SendingSmartAppNotification
 from botx.models.status import StatusRecipient
 from botx.testing.building.builder import MessageBuilder
 
@@ -128,6 +129,8 @@ __all__ = (
     "NotificationOptions",
     "MessagePayload",
     "UpdatePayload",
+    "SendingSmartAppEvent",
+    "SendingSmartAppNotification",
     # testing
     "TestClient",
     "MessageBuilder",