+++ botx/testing/botx_mock/wsgi/application.py
@@ -12,6 +12,7 @@ from botx.testing.botx_mock.wsgi.routes import (
     command,
     events,
     notification,
+    smartapps,
     users,
 )
 from botx.testing.typing import APIMessage, APIRequest
@@ -41,6 +42,9 @@ _ENDPOINTS: Tuple[Callable[..., Any], ...] = (
     users.get_by_huid,
     users.get_by_email,
     users.get_by_login,
+    # smartapps
+    smartapps.post_smartapp_event,
+    smartapps.post_smartapp_notification,
 )