+++ tests/test_collecting/test_collector/test_decorators/test_system_event.py
@@ -16,6 +16,7 @@ def test_registration_handler_for_several_system_events(
         SystemEvents.added_to_chat,
         SystemEvents.deleted_from_chat,
         SystemEvents.left_from_chat,
+        SystemEvents.smartapp_event,
     }
     collector = collector_cls()
     collector.system_event(
@@ -34,6 +35,7 @@ def test_registration_handler_for_several_system_events(
         SystemEvents.chat_created,
         SystemEvents.file_transfer,
         SystemEvents.left_from_chat,
+        SystemEvents.smartapp_event,
     ],
 )
 def test_defining_system_handler_in_collector_as_decorator(