+++ docs/changelog.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Added
+
+* Add smartapps main functional.
+
+
 ## 0.21.2 (Aug 3, 2021)
 
 ### Fixed